Fix test for CreateAppStepWithDocker not executed
