+++ com.sap.cloud.lm.sl.cf.process/src/test/java/com/sap/cloud/lm/sl/cf/process/steps/CreateAppStepBaseTest.java
@@ -1,0 +1,39 @@
+package com.sap.cloud.lm.sl.cf.process.steps;
+
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.cloudfoundry.client.lib.domain.ServiceKey;
+
+import com.sap.cloud.lm.sl.cf.client.lib.domain.CloudApplicationExtended;
+import com.sap.cloud.lm.sl.cf.client.lib.domain.CloudServiceExtended;
+import com.sap.cloud.lm.sl.cf.core.cf.PlatformType;
+
+public abstract class CreateAppStepBaseTest extends SyncFlowableStepTest<CreateAppStep> {
+
+    protected StepInput stepInput;
+    protected CloudApplicationExtended application;
+    
+    protected static class StepInput {
+        List<CloudApplicationExtended> applications = Collections.emptyList();
+        List<SimpleService> services = Collections.emptyList();
+        int applicationIndex;
+        PlatformType platform;
+        Map<String, String> bindingErrors = new HashMap<>();
+        Map<String, List<ServiceKey>> existingServiceKeys = new HashMap<>();
+    }
+
+    protected static class SimpleService {
+        String name;
+        boolean isOptional;
+
+        CloudServiceExtended toCloudServiceExtended() {
+            CloudServiceExtended service = new CloudServiceExtended();
+            service.setName(name);
+            service.setOptional(isOptional);
+            return service;
+        }
+    }
+}
